+++ _gulp/server.js
@@ -1,0 +1,32 @@
+const Server = require('spa-server');
+const path = require('path');
+const url = require('url');
+const shell = require('shelljs');
+
+const spawnServer = (folder) => {
+  const targetFolder = path.resolve(folder, '_site');
+  const server = Server.create({
+    path: '_site',
+    port: 8888,
+    fallback: (req, res) => {
+      const parsed = url.parse(req.url),
+            target = parsed.pathname + '.html',
+            alternateTarget = parsed.pathname + '/index.html';
+
+      console.log(path.join(targetFolder, target));
+
+      if (shell.test('-f', path.resolve(targetFolder, target))) {
+        return target + (parsed.query ? `?${parsed.query}` : '');
+      }
+      if (shell.test('-f', path.resolve(targetFolder, alternateTarget))) {
+        return alternateTarget + (parsed.query ? `?${parsed.query}` : '');
+      }
+      return '404.html';
+    }
+  })
+  server.start();
+};
+
+module.exports = {
+  spawn: spawnServer
+};
